Updates because of new packer version
It seems you now must define a checksum for ova/ovf input files. It also
seems that the when builds picked up the output of the base box that
they were converting the ovf to an ova. The ovf + the new checksum stuff
wasn't working right anyhow so I switched to ova and everthing started
working.

The last change was that I dropped ruby 1.9.3 from the rvm image due to
it failing to build.